requirements.txt
opencv-python==4.5.3.56
pytesseract==0.3.8
mss==6.1.0
pillow==8.2.0
matplotlib==3.4.3
pyautogui==0.9.53
pynput==1.7.3